Move Lock to private namespace, add interface. Update PHPDoc.
